Refactor GP finals assigned cup validator
